feat: tooltip optimize, related #427
